+++ release-notes.md
@@ -1,5 +1,8 @@
 Lemur 1.14.1 (unreleased)
 -------------
+* Fixed a bug where moving a GuiComponent from one GUI element to another
+    caused lingering state issues.  This most prominently manifested when
+    using icons in ListBox labels.
 
 
 Lemur 1.14.0 (latest)